chore: moving settings and report-hazard inside pages folder
