+++ pages/report-hazard/style.css
@@ -1,0 +1,4 @@
+#map {
+  height: 400px;
+  width: 100vw;
+}
