+++ pages/settings/index.html
@@ -1,0 +1,91 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+  <!-- WPA -->
+  <link rel="manifest" href="../../manifest.json">
+  <link rel="apple-touch-icon" href="../../assets/css/img/icons/icon-96x96.png">
+  <meta name="apple-mobile-web-app-status-bar" content="#000000">
+  <meta name="theme-color" content="#000000">
+
+  <!-- Bootstrap JS -->
+  <script src="../../modules/bootstrap/js/bootstrap.min.js"></script>
+  <!-- Bootstrap CSS -->
+  <link rel="stylesheet" href="../../modules/bootstrap/css/bootstrap.min.css">
+  <!-- Variables -->
+  <link rel="stylesheet" href="../../assets/css/variables.css">
+  <!-- Global Styles -->
+  <link rel="stylesheet" href="../../assets/css/global.css">
+  <!-- Feature Styles -->
+  <link rel="stylesheet" href="./style.css">
+  <!-- Global JS -->
+  <script type="module" src="../../global.js" defer></script>
+  <!-- Feature JS -->
+  <script type="module" src="./script.js" defer></script>
+
+  <title>Settings</title>
+</head>
+
+<body>
+  <section id="myProfile">
+    <h1>My Profile</h1>
+    <p>Update your photo and personal details here.</p>
+    <form id="profileForm">
+      <label for="name">Name</label>
+      <input type="text" id="name" />
+      <label for="email">Email address</label>
+      <input type="email" id="email" />
+    </form>
+    <h2>Password</h2>
+    <button id="changePwBtn">Change Password</button>
+    <h2>Your photo</h2>
+    <p>This will be displayed on your profile.</p>
+    <img src="#" alt="#" id="profilePhoto" />
+    <!-- change label css to look like button -->
+    <label for="changeProfilePhotoBtn">Change profile photo</label>
+    <!-- hide input field in css -->
+    <input id="changeProfilePhotoBtn" type="file" accept="image/jpeg, image/jpg, image/png" />
+
+    <button type="submit" id="saveProfileInfoBtn" form="profileForm">Save</button>
+  </section>
+  <section id="settings">
+    <h1>Settings</h1>
+    <p>Update your preferences here</p>
+
+    <div class="form-check form-switch">
+      <img src="#" alt="#" />
+      <!-- icon -->
+      <label class="form-check-label" for="pushNotificationSwitch">Push Notifications</label>
+      <input class="form-check-input" type="checkbox" role="switch" id="pushNotificationSwitch" checked />
+    </div>
+    <div id="changePwPopup">
+      <form id="changePwForm">
+        <label for="oldPw">Old Password</label>
+        <input type="password" id="oldPw" />
+        <label for="newPw">New Password</label>
+        <input type="password" id="newPw" />
+        <label for="newPwConfirm">Confirm New Password</label>
+        <input type="password" id="newPwConfirm" />
+        <button id="changePwSaveBtn">Save & Close</button>
+      </form>
+    </div>
+  </section>
+  <section id="deleteAccount">
+    <h1>Delete Account</h1>
+    <p>
+      If you no longer want to use CampBuddy, you can permanently delete your
+      account. You can’t undo this action!
+    </p>
+    <button id="deleteAccountBtn">Delete Account</button>
+    <div id="deleteAccountConfirm">
+      <h2>Are you sure you want to delete your account?</h2>
+      <button id="deleteAccountYesBtn">Yes</button>
+      <button id="deleteAccountNoBtn">No</button>
+    </div>
+  </section>
+</body>
+
+</html>
