--- report-hazard/style.css
@@ -1,4 +1,0 @@
-#map {
-  height: 200px;
-  width: 200px;
-}
